I had 2 issues with PM2 :
The script generated by pm2 startup was using the wrong user (root instead of pi)
The dump file containing the applications to ressurect was set to empty by PM2 when rebooting, therefore the applications where not starting automatically when rebooting manually
So I made 2 changes to rpi-2-install.sh script.

This might need to be tested on distro others than raspbian.